Add dinosaur theme
This adds the html, css, and js required for a cool theme, inspired by
the winston theme.

I'll continue to iterate on this as I write more posts and see what
works, as I'd like to have my old blog header back in too, as I
think that was pretty cool.