Add time to read for blog posts
This adds the time it takes to read the blog post everywhere we list the
author of the post.

It works out the time taken to read based on an average of
250 words read per minute, then rounded up to ensure we don't have a
number less than 1. This is the midway point between the average 200-300
words per minute for general reading speed.

The author is missing from the tag page, but I will add it there at
some point.